FontMan: Change type sizes at glyph structures
This should optimise the memory use

+++ src/fontman/raster_font.cpp
@@ -237,8 +237,8 @@ void RasterFont::loadFontMap(const std::string& fontmap_ini)
             rch.padding_left    = (ucharX.size() > 1) ? char2int(ucharX[1]) : 0;
             rch.padding_right   = (ucharX.size() > 2) ? char2int(ucharX[2]) : 0;
             t                   =  std::stof(charPosX.c_str()) / m_matrixHeight;
-            rch.x               =  static_cast<int32_t>(fontTexture.w * l);
-            rch.y               =  static_cast<int32_t>(fontTexture.h * t);
+            rch.x               =  static_cast<int16_t>(fontTexture.w * l);
+            rch.y               =  static_cast<int16_t>(fontTexture.h * t);
             rch.valid = true;
         }
         catch(std::exception &e)
